Initial Commit
This has a single .py file that holds the start of this utility. In its current form it is able to track the start and lengths of straight from the driver's throttle inputs. From that it finds how much of the avaible hybrid deploy per lap (hard coded to 100 right now) to deploy at that location on track.