Radical Change
Realized that it is much easier to classify based on whether manual boost should be active or not. So exported a motec log file from sebring to a .csv file to use as a small test data set. Then used sklearn's prebuilt models to test different classification models. So far SVM (SVC) has worked best on the, admittedly small, dataset.